# api/openapi.yaml
openapi: 3.0.3
info:
  title: Nextstep Dashboard API
  version: 1.1.0
  description: >-
    Contract for the multi-tenant SML dashboard, production administration,
    distributed report execution, LINE delivery, and LIFF viewer. Mutation
    requests authenticated by cookie also require X-CSRF-Token.
servers:
  - url: https://dashboard.nextstep-soft.com
...
paths:
...
  /api/v1/admin/tenants:
    get:
      tags: [Admin tenants]
      operationId: listTenants
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/PageSize'}
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/TenantStatus'}
        - name: search
          in: query
          schema: {type: string, maxLength: 160}
      responses:
        '200':
          description: Cursor-paginated tenants.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TenantPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      tags: [Admin tenants]
      operationId: createTenant
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TenantInput'}
      responses:
        '201':
          description: Tenant created disabled by default.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Tenant'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/v1/viewer/line/session:
    post:
      tags: [Viewer]
      operationId: exchangeLineSession
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LineSessionInput'}
      responses:
        '200':
          description: LINE identity verified and viewer cookie set.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ViewerMe'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /api/v1/viewer/logout:
    post:
      tags: [Viewer]
      operationId: logoutViewer
      security: [{ViewerSession: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Viewer session revoked.}
  /api/v1/viewer/me:
    get:
      tags: [Viewer]
      operationId: getViewerMe
      security: [{ViewerSession: []}]
      responses:
        '200':
          description: Current verified viewer.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ViewerMe'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/viewer/tenants:
    get:
      tags: [Viewer]
      operationId: listViewerTenants
      security: [{ViewerSession: []}]
      responses:
        '200':
          description: Active, unexpired tenant memberships.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/ViewerTenant'}}
  /api/v1/viewer/tenants/{tenantId}/reports:
    get:
      tags: [Viewer]
      operationId: listViewerReports
      security: [{ViewerSession: []}]
      parameters: [{$ref: '#/components/parameters/TenantID'}]
      responses:
        '200':
          description: Permission-filtered report catalog.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/ReportDefinition'}}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: nextstep_admin_session
    ViewerSession:
      type: apiKey
      in: cookie
      name: nextstep_viewer_session
  parameters:
    TenantID: {name: tenantId, in: path, required: true, schema: {type: string, format: uuid}}
...
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 512}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
    CSRFToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 32, maxLength: 256}}
  responses:
    Unauthorized: {$ref: '#/components/responses/Problem401'}
    Forbidden: {$ref: '#/components/responses/Problem403'}
...
    Conflict: {$ref: '#/components/responses/Problem409'}
...
    ValidationFailed: {$ref: '#/components/responses/Problem422'}
...
    Problem401: {description: Authentication required or invalid., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
    Problem403: {description: Authenticated identity is not authorized., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
    Problem409: {description: 'Duplicate request, state conflict, or optimistic lock conflict.', content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
    Problem422: {description: Input is syntactically valid but violates a business rule., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
  schemas:
...
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            retryable: {type: boolean}
            fieldErrors: {type: array, items: {$ref: '#/components/schemas/FieldError'}}
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: {type: string}
        code: {type: string}
        message: {type: string}
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: string, nullable: true}
        hasMore: {type: boolean}
...
    TenantStatus: {type: string, enum: [ACTIVE, DISABLED, EXPIRED]}
    Tenant:
      type: object
      required: [id, slug, name, timezone, status, accessEndsAt, version, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        timezone: {type: string, enum: [Asia/Bangkok]}
        status: {$ref: '#/components/schemas/TenantStatus'}
        accessEndsAt: {type: string, format: date-time}
        version: {type: integer, minimum: 1}
        smlReadiness: {type: string, enum: [UNCONFIGURED, UNTESTED, READY, FAILED]}
        nextScheduleAt: {type: string, format: date-time, nullable: true}
        viewerUrl: {type: string, format: uri, readOnly: true, description: Canonical Dashboard link generated from configured PUBLIC_BASE_URL; opening it never grants access.}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    TenantInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        slug: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 80, deprecated: true, description: Compatibility field for older clients; omitted by the current admin UI.}
        name: {type: string, minLength: 1, maxLength: 160}
        timezone: {type: string, enum: [Asia/Bangkok], deprecated: true}
        accessEndsAt: {type: string, format: date-time, deprecated: true}
...
    TenantPage:
      type: object
      required: [data, page]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/Tenant'}}
        page: {$ref: '#/components/schemas/PageInfo'}
...
    ReportKey:
      type: string
      enum: [sales_goods_services, purchase_goods_payables, gross_profit_by_product, gross_profit_by_ar_customer, stock_balance, stock_reorder, ar_customer_movement, ar_debt_receipt, cash_bank_receipts, cash_bank_payments]
    ReportDefinition:
      type: object
      required: [reportKey, version, label, category, isSensitive]
      properties:
        reportKey: {$ref: '#/components/schemas/ReportKey'}
        version: {type: string}
        label: {type: string}
        category: {type: string}
        isSensitive: {type: boolean}
        periodMode: {type: string, enum: [DATE_RANGE, AS_OF_DATE, CURRENT_ONLY]}
...
    ReportPeriod:
      type: object
      additionalProperties: false
      required: [preset, dateFrom, dateTo]
      properties:
        preset: {type: string, enum: [YESTERDAY, TODAY_TO_NOW, MONTH_TO_DATE, AS_OF_RUN, CUSTOM]}
        dateFrom: {type: string, format: date}
        dateTo: {type: string, format: date}
    MetricUnit:
      type: string
      enum: [THB, COUNT, PERCENT, QUANTITY, RATIO]
    MetricComparison:
      type: object
      additionalProperties: false
      required: [availability]
      properties:
        availability: {type: string, enum: [AVAILABLE, UNAVAILABLE]}
        previousValue: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        delta: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        percent: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        direction: {type: string, enum: [UP, DOWN, SAME]}
    DashboardMetric:
      type: object
      additionalProperties: false
      required: [key, label, value, unit, comparison]
      properties:
        key: {type: string, minLength: 1, maxLength: 80}
        label: {type: string, minLength: 1, maxLength: 160}
        value: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        unit: {$ref: '#/components/schemas/MetricUnit'}
        comparison: {$ref: '#/components/schemas/MetricComparison'}
    VisualizationSeries:
      type: object
      additionalProperties: false
      required: [key, label, values]
      properties:
        key: {type: string, minLength: 1, maxLength: 80}
        label: {type: string, minLength: 1, maxLength: 160}
        values:
          type: array
          maxItems: 92
          items: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        pointLabels:
          type: array
          maxItems: 92
          items: {type: string, maxLength: 80}
    DashboardVisualization:
      type: object
      additionalProperties: false
      required: [key, title, intent, unit, categories, series]
      properties:
        key: {type: string, minLength: 1, maxLength: 80}
        title: {type: string, minLength: 1, maxLength: 160}
        intent: {type: string, enum: [TREND, RANKING, COMPOSITION, EXCEPTION]}
        unit: {$ref: '#/components/schemas/MetricUnit'}
        categories:
          type: array
          maxItems: 92
          items: {type: string, maxLength: 160}
        series:
          type: array
          minItems: 1
          maxItems: 6
          items: {$ref: '#/components/schemas/VisualizationSeries'}
        note: {type: string, maxLength: 500}
    DashboardQuality:
      type: object
      additionalProperties: false
      required: [status, warnings]
      properties:
        status: {type: string, enum: [OK, WARNING]}
        warnings:
          type: array
          maxItems: 20
          items: {type: string, maxLength: 120}
    ReportDashboard:
      type: object
      additionalProperties: false
      required: [reportKey, version, period, comparisonPeriod, timezone, generatedAt, kpis, visualizations, quality]
      properties:
        reportKey: {$ref: '#/components/schemas/ReportKey'}
        version: {type: string, minLength: 1, maxLength: 32}
        period: {$ref: '#/components/schemas/ReportPeriod'}
        comparisonPeriod: {$ref: '#/components/schemas/ReportPeriod'}
        timezone: {type: string, enum: [Asia/Bangkok]}
        generatedAt: {type: string, format: date-time}
        kpis:
          type: array
          minItems: 2
          maxItems: 8
          items: {$ref: '#/components/schemas/DashboardMetric'}
        visualizations:
          type: array
          maxItems: 4
          items: {$ref: '#/components/schemas/DashboardVisualization'}
        quality: {$ref: '#/components/schemas/DashboardQuality'}
    DashboardSnapshot:
      type: object
      additionalProperties: false
      required: [runId, dashboard]
      properties:
        runId: {type: string, format: uuid}
        dashboard: {$ref: '#/components/schemas/ReportDashboard'}
        periodFrom: {type: string, format: date}
        periodTo: {type: string, format: date}
        sourceStartedAt: {type: string, format: date-time, nullable: true}
        sourceFinishedAt: {type: string, format: date-time, nullable: true}
        freshUntil: {type: string, format: date-time, nullable: true}
        staleUntil: {type: string, format: date-time, nullable: true}
        freshnessStatus: {type: string, enum: [FRESH, STALE, EXPIRED, MISSING, REFRESHING, REFRESH_FAILED]}
        reportDefinitionVersion: {type: string}
        dataSourceVersion: {type: integer, minimum: 0}
        queryPlanFingerprint: {type: string, pattern: '^[0-9a-f]{64}$'}
        sourceConsistency: {type: string, enum: [STATEMENT, SERIAL_WINDOW, CHUNK_WINDOW]}
        detailsAvailable: {type: boolean}
        detailsExpiresAt: {type: string, format: date-time, nullable: true}
...
    LineSessionInput:
      type: object
      additionalProperties: false
      required: [idToken]
      properties:
        idToken: {type: string, minLength: 32, maxLength: 8192, writeOnly: true}
        invitationReference: {type: string, minLength: 32, maxLength: 128, writeOnly: true}
        deliveryReference: {type: string, minLength: 32, maxLength: 512, writeOnly: true}
        expectedTenantId: {type: string, format: uuid, description: Explicit tenant from the viewer route; mismatch is rejected without redirecting to another tenant.}
    ViewerMe:
      type: object
      required: [recipientId, displayName, expiresAt]
      properties:
        recipientId: {type: string, format: uuid}
        displayName: {type: string}
        csrfToken: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
        deliveryContext: {$ref: '#/components/schemas/DeliveryContext'}
        deliveryContextErrorCode: {type: string, enum: [DELIVERY_CONTEXT_UNAVAILABLE, DELIVERY_CONTEXT_PERMISSION_CHANGED]}
    ViewerTenant:
      type: object
      required: [id, name, timezone, reportKeys]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        timezone: {type: string}
        reportKeys: {type: array, uniqueItems: true, items: {$ref: '#/components/schemas/ReportKey'}}
...
    DeliverySnapshotStatus:
      type: string
      enum: [AVAILABLE, DETAIL_EXPIRED, SNAPSHOT_EXPIRED, UNAVAILABLE]
    DeliveryContextReport:
      type: object
      required: [reportKey, label, reportRunId, snapshotStatus]
      properties:
        reportKey: {$ref: '#/components/schemas/ReportKey'}
        label: {type: string}
        position: {type: integer, minimum: 1, maximum: 10}
        reportRunId: {type: string, format: uuid}
        snapshotStatus: {$ref: '#/components/schemas/DeliverySnapshotStatus'}
        summary: {$ref: '#/components/schemas/DashboardSnapshot'}
    DeliveryContext:
      type: object
      additionalProperties: false
      required: [deliveryId, tenantId, scheduledFor, materializationVersion, orderStatus, dataStatus, reports]
      properties:
        deliveryId: {type: string, format: uuid}
        tenantId: {type: string, format: uuid}
        scheduledFor: {type: string, format: date-time}
        materializationVersion: {type: integer, minimum: 1}
        orderStatus: {type: string, enum: [EXACT, LEGACY]}
        dataStatus: {type: string, enum: [AVAILABLE, PARTIAL_EXPIRED, EXPIRED]}
        reports: {type: array, minItems: 1, maxItems: 10, items: {$ref: '#/components/schemas/DeliveryContextReport'}}
...